pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "leadenrich-backend"
version = "0.1.0"
description = "Lead enrichment platform backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.0",
    "sqlalchemy==2.0.35",
    "psycopg2-binary==2.9.9",
    "zhipuai==2.1.5",
    "firecrawl-py==1.6.0",
    "python-dotenv==1.0.1",
    "httpx==0.27.0",
    "pydantic==2.9.0",
    "pydantic-settings==2.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "black==23.12.1",
    "ruff==0.1.8",
]
